% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_-.R
\name{-,Dyadic,Dyadic-method}
\alias{-,Dyadic,Dyadic-method}
\title{Subtraction of dyadic objects}
\usage{
\S4method{-}{Dyadic,Dyadic}(e1, e2)
}
\arguments{
\item{e1}{A \code{Dyadic}-object.}

\item{e2}{A \code{Dyadic}-object.}
}
\value{
The \code{Dyadic}-object that is the result of the subtraction.
}
\description{
The subtraction of dyadic objects.
}
\details{
The operations are performed in a way that is consistent with the dyadic structure of the matrices.
}
\section{References}{

Kos, M., Podg\eqn{\mbox{\'o}}{o}rski, K., Wu, H. (2024) "Sparse"
}

\examples{
#----------------------------------------------#
#------- Subtraction of dyadic objects --------#
#----------------------------------------------#

N <- 4
k <- 3

# Construct four types of dyadic matrices with made of 1's
V <- construct(N, k, type = "vert") # vertical
H <- construct(N, k, type = "horiz") # horizontal
S <- construct(N, k, type = "symm") # symmetric
AS <- construct(N, k, type = "asymm") # asymmetric

# Convert the dyadic matrices to matrix format
mat_V <- as.matrix(V)
mat_H <- as.matrix(H)
mat_S <- as.matrix(S)
mat_AS <- as.matrix(AS)

# Subtract two dyadic matrices

HpAS <- 3 * t(H) - 6 * t(AS) # transpose and linear combination
HpAS@type # check the type of the result
}
\seealso{
\code{\link{Dyadic-class}} for the definition of the \code{Dyadic}-class;
\code{\link{dyadFac}} for the dyadic decomposition of dyadic matrices;
}
