% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_neg.R
\name{-,Dyadic-method}
\alias{-,Dyadic-method}
\alias{-,Dyadic-method,}
\alias{-,Dyadic,ANY-method}
\title{Negation of dyadic objects}
\arguments{
\item{e1}{A \code{Dyadic}-object.}
}
\value{
The \code{Dyadic}-object with negated entries and aentries.
}
\description{
Negates the entries and aentries of a dyadic object.
}
\details{
The negation is performed in a way that is consistent with
the dyadic structure of the matrices.
}
\section{References}{

Kos, M., Podg\eqn{\mbox{\'o}}{o}rski, K., Wu, H. (2024) "Sparse"
}

\examples{
#----------------------------------------------#
#--------- Negation of dyadic objects ---------#
#----------------------------------------------#

N <- 4
k <- 3

# Construct four types of dyadic matrices with made of 1's
V <- construct(N, k, type = "vert") # vertical
NV <- -V

mat_V <- as.matrix(V)
mat_NV <- as.matrix(NV)

max(abs(mat_V + mat_NV)) == 0 # Should be 0
}
\seealso{
\code{\link{Dyadic-class}} for the definition of the \code{Dyadic}-class;
\code{\link{dyadFac}} for the dyadic decomposition of dyadic matrices;
}
