% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_t.R
\name{t,Dyadic-method}
\alias{t,Dyadic-method}
\title{Construction of a \code{Dyadic} object}
\usage{
\S4method{t}{Dyadic}(x)
}
\arguments{
\item{x}{\code{Dyadic}-object;}
}
\value{
The \code{Dyadic}-object that is the result of the operation with properly defined fields.
}
\description{
The \code{Dyadic} object transpose of a \code{Dyadic} object: \code{(t(Dyadic)}.
}
\details{
The operations are performed in a way that is consistent with the dyadic structure of the matrices.
}
\section{References}{

Kos, M., Podg\eqn{\mbox{\'o}}{o}rski, K., Wu, H. (2024) "Sparse"
}

\examples{
#--------------------------------------------#
#-------Transpose of a dyadic object --------#
#--------------------------------------------#

N <- 4
k <- 3

# Construct four types of dyadic matrices with made of 1's
V <- construct(N, k, type = "vert") # vertical
H <- construct(N, k, type = "horiz") # horizontal
S <- construct(N, k, type = "symm", distr = "unif") # symmetric


V@type
H@type
t(H)@type
as.matrix(S)
as.matrix(t(S))
}
\seealso{
\code{\link{Dyadic-class}} for the definition of the \code{Dyadic}-class;
\code{\link{dyadFac}} for the dyadic decomposition of dyadic matrices;
}
