% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_*.R
\name{*,Dyadic,numeric-method}
\alias{*,Dyadic,numeric-method}
\title{Scalar multiplication of dyadic objects}
\usage{
\S4method{*}{Dyadic,numeric}(e1, e2)
}
\arguments{
\item{e1}{numeric scalar.}

\item{e2}{\code{Dyadic}-object.}
}
\value{
The \code{Dyadic}-object that is the result of the operation.
}
\description{
The scalar multiplication of a dyadic object.
}
\details{
The operations are performed in a way that is consistent with the dyadic structure of the matrices.
}
\section{References}{

Kos, M., Podg\eqn{\mbox{\'o}}{o}rski, K., Wu, H. (2024) "Sparse"
}

\examples{
#--------------------------------------------------------#
#------- Scalar multiplication of dyadic objects --------#
#--------------------------------------------------------#

# Construct four types of dyadic matrices with made of 1's
V <- construct(N, k, type = "vert") # vertical

mat_V <- as.matrix(V)

as.matrix(2*V)==2*mat_V

}
\seealso{
\code{\link{Dyadic-class}} for the definition of the \code{Dyadic}-class;
\code{\link{dyadalg}} for the dyadic decomposition of dyadic matrices;
}
